Add files via upload
It is a website that provides information about the rooms available in your locality, based on the location and your very own preferences. It has the most frequent feature of sorting rooms based on the preference you give and showing them on the map. No more hustle and chaos, just select and click, and your room will be booked. You can also ask and share, right now it is in underdevelopment so until then enjoy this prototype!